Minor fixes, see detailed log msg however
The text() is underlying the rectangle on which the GLSL shaders are
rendered. For the URL, we can add a div and place the URL on the bottom
of the screen, or just bypass it altogether (not recommended though).
For the players information however, it seems we will need to overlap
some information in the GLSL shader itself, since we cannot draw it in
the p5js sketch. The GLSL rectangle has precedence...